move otaclient._utils.logging into otaclient_common package
